Tests to make sure that Sockets implement all Stream methods and events
